use u8 for bools

+++ src/decode_bmp.h
@@ -25,14 +25,14 @@ void get_BMP_width_height(
     const uint64_t raw_input_size,
     uint32_t * out_width,
     uint32_t * out_height,
-    uint32_t * out_good);
+    uint8_t * out_good);
 
 void decode_BMP(
     const uint8_t * raw_input,
     const uint64_t raw_input_size,
     uint8_t * out_rgba_values,
     const int64_t out_rgba_values_size,
-    uint32_t * out_good);
+    uint8_t * out_good);
 
 void encode_BMP(
     const uint8_t * rgba,
